Промежуток

+++ src/main/java/pl/pionwit/models/interfaces/AddressImpl.java
@@ -1,11 +1,11 @@
 package pl.pionwit.models.interfaces;
 
 
-import pl.pionwit.models.AddressesEntity;
+import pl.pionwit.dbmain.dbtables.AddressesEntity;
 
 public interface AddressImpl {
     void addUpdate(AddressesEntity address);
     void dell (AddressesEntity address);
-    AddressesEntity search (int id);
+    AddressesEntity search (Long id);
     boolean isSet (AddressesEntity address);
 }
