+++ src/main/java/pl/pionwit/models/interfaces/TypesAddressImpl.java
@@ -1,11 +1,11 @@
 package pl.pionwit.models.interfaces;
 
 
-import pl.pionwit.models.TypesAddressEntity;
+import pl.pionwit.dbmain.dbtables.TypesAddressEntity;
 
 public interface TypesAddressImpl {
     void addUpdate(TypesAddressEntity typeAddress);
     void dell (TypesAddressEntity typeAddress);
-    TypesAddressEntity search (int id);
+    TypesAddressEntity search (Long id);
     boolean isSet (TypesAddressEntity typeAddress);
 }
